% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{cal_log2_ratio}
\alias{cal_log2_ratio}
\title{Calculate log2 ratio of observed and expected values of intensities.}
\usage{
cal_log2_ratio(query, reference, method = c("mean", "median", "lm"))
}
\arguments{
\item{query}{A matrix of query sample intensity.}

\item{reference}{A matrix of reference sample intensity.}

\item{method}{A character scalar. Default to "mean".}
}
\value{
A matrix of log2 ratio of query samples.
}
\description{
Calculate log2 ratio of observed and expected values of intensities at each
probe.
}
\note{
If an intensity is less than 1, it is set to 1.
}
