% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conumee-annotation.R
\name{CNV.create_anno.yamat}
\alias{CNV.create_anno.yamat}
\title{Create CNV annotaion with \code{yamat} preset.}
\usage{
CNV.create_anno.yamat(x)
}
\arguments{
\item{x}{An object of \code{\link[minfi]{RGChannelSet-class}} or
\code{\link[minfi]{MethylSet-class}} or
\code{\link[minfi]{GenomicMethylSet-class}}.}
}
\value{
An object of \code{\link[conumee]{CNV.anno-class}}.
}
\description{
The function is a wrapper of \code{\link{CNV.create_anno2}} with
pre-defined settings of detail regions.
}
\details{
Exclude regions are from \code{conumee} package.
}
