% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conumee-annotation.R
\name{CNV.create_anno2}
\alias{CNV.create_anno2}
\title{Create Conumee annotation for CNV analysis.}
\usage{
CNV.create_anno2(array_type = "EPIC", array_anno = "ilm10b4.hg19",
  bin_minprobes = 15, bin_minsize = 50000, bin_maxsize = 5e+06,
  chrXY = FALSE, exclude_regions = NULL, detail_regions = NULL)
}
\arguments{
\item{array_type}{A character scalar. One of "450k", "EPIC", or "overlap".
Defaults to "EPIC".}

\item{array_anno}{A character scalar. One of "ilm10b2.hg19", "ilm10b3.hg19",
"ilm10b4.hg19". Only valid when \code{array_type} is set to "EPIC". Default
to "ilm10b4.hg19".}

\item{bin_minprobes}{numeric. Minimum number of probes per bin. Bins
are interatively merged with neighboring bin until minimum number is
reached.}

\item{bin_minsize}{numeric. Minimum size of a bin.}

\item{bin_maxsize}{numeric. Maximum size of a bin. Merged bins that
are larger are filtered out.}

\item{chrXY}{logical. Should chromosome X and Y be included in the analysis?}

\item{exclude_regions}{GRanges object or path to bed file containing
genomic regions to be excluded.}

\item{detail_regions}{GRanges object or path to bed file containing
genomic regions to be examined in detail.}
}
\value{
An object of \code{\link[conumee]{CNV.anno-class}}.
}
\description{
Rewrite \code{\link[conumee]{CNV.create_anno}()} to be compatible with
ilm10b3.hg19 and ilm10b4.hg19 annotation of EPIC array. I also change the
default array type to EPIC.
}
\details{
This function collects all annotations required for CNV analysis
using Illumina 450k or EPIC/850K arrays. I rewrite the function so that
it can generate the annotation for the list of probes specified by the
experiments.
}
\note{
I have tested with EPIC ilm10b4.hg19, but not other arrays and
  annotations.
}
