% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conumee-pipe.R
\name{conumee_backbone}
\alias{conumee_backbone}
\title{CNV pipeline backbone.}
\usage{
conumee_backbone(x, report_dir, pipe_file = "conumee.Rda",
  dnacopy_seed = 1, overwrite = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{x}{An object of \code{\link[minfi]{GenomicRatioSet-class}} or
\code{\link[minfi]{GenomicMethylSet-class}}.}

\item{report_dir}{A character scalar of reporting.}

\item{pipe_file}{A character scalar of the filename storing the pipeline
result sample by sample. Default to "conumee.Rda". Turn off the saving by
set it to NULL.}

\item{dnacopy_seed}{An integer scalar to set the seed. Default to 1.}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}

\item{...}{Any arguments passed to \code{\link[conumee]{CNV.segment}}.}
}
\value{
An object of \code{\link{ConumeePipe}} class.
}
\description{
CNV pipeline backbone is the steps after preprocessing, including loading
data into Conumee, analysis and write the result.
}
\details{
The \code{ref} and \code{qry} should include at least the following
  columns \code{Sample_Name}, \code{Gender}, \code{Batch}, \code{Sample_Prep}.
  Batch effects are removed on methylation and unmethylation signals (log2
  transformed) separately.
}
