% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAcopy.R
\name{create_dnacopy_cna}
\alias{create_dnacopy_cna}
\title{Create \code{\link[DNAcopy]{CNA}} object.}
\usage{
create_dnacopy_cna(x, lrr, verbose = TRUE)
}
\arguments{
\item{x}{A \code{minfi} object such as
\code{\link[minfi]{RGChannelSet-class}} or
\code{\link[minfi]{MethylSet-class}} or
\code{\link[minfi]{GenomicMethylSet-class}} or
\code{\link[minfi]{GenomicRatioSet-class}}.}

\item{lrr}{A matrix of log2 transformed ratio of query and reference copy
number (sum of methylation and unmethylation signals).}

\item{verbose}{A logical scalar. Default to TRUE.}
}
\value{
A \code{\link[DNAcopy]{CNA}} object.
}
\description{
Create \code{\link[DNAcopy]{CNA}} object.
}
