% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwob-pipe.R
\name{cwob_pipe}
\alias{cwob_pipe}
\title{CWOB pipeline.}
\usage{
cwob_pipe(ref, qry, report_dir, norm_method = c("swan", "methylcnv",
  "illumina", "raw", "quantile", "noob", "funnorm", "yamat", "dkfz"),
  batch = NULL, batch2 = NULL, pipe_file = "cwob.Rda",
  dnacopy_seed = 1, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{ref}{Reference samples stored in an object of
\code{\link[minfi]{RGChannelSet-class}}.}

\item{qry}{Query samples stored in an object of
\code{\link[minfi]{RGChannelSet-class}}.}

\item{report_dir}{A character scalar of reporting.}

\item{norm_method}{A character scalar of method, including raw, illumina,
swan, quantile, noob, funnorm, yamat, dkfz, methylcnv. Default to
"swan".}

\item{batch}{factor or vector indicating batches. Default to \code{NULL}, -
do not remove batch effect.}

\item{batch2}{optional factor or vector indicating a second series of
batches. Default to \code{NULL}, - do not remove batch effect.}

\item{pipe_file}{A character scalar of the filename storing the pipeline
result sample by sample. Default to "cwob.Rda". Turn off the saving by
set it to NULL.}

\item{dnacopy_seed}{An integer scalar to set the seed. Default to 1.}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}
}
\value{
An object of \code{\link{CwobPipe}} class.
}
\description{
Conumee pipeline without binning.
}
\details{
The \code{ref} and \code{qry} should include at least the following
  columns \code{Sample_Name}, \code{Gender}, \code{Batch}, \code{Sample_Prep}.
  Batch effects are removed on methylation and unmethylation signals (log2
  transformed) separately.
}
