% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conumee-annotation.R
\name{detail_regions}
\alias{detail_regions}
\title{Get detail regions: yamat}
\usage{
detail_regions(thick = 5e+05)
}
\arguments{
\item{thick}{An integer scalar defined the flanking region size in bp.
Default to 500000.}
}
\value{
A \code{GRange} object.
}
\description{
The detail regions are used in the authors' studies, focusing on pediatric
brain tumors and solid tumors.
}
