% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAcopy.R
\name{dnacopy_analysis.conumee}
\alias{dnacopy_analysis.conumee}
\title{Run DNAcopy analysis: Conumee's settings.}
\usage{
dnacopy_analysis.conumee(x, lrr, seed = 1, verbose = TRUE)
}
\arguments{
\item{x}{A \code{minfi} object such as
\code{\link[minfi]{RGChannelSet-class}} or
\code{\link[minfi]{MethylSet-class}} or
\code{\link[minfi]{GenomicMethylSet-class}} or
\code{\link[minfi]{GenomicRatioSet-class}}.}

\item{lrr}{A matrix of log2 transformed ratio of query and reference copy
number (sum of methylation and unmethylation signals).}

\item{seed}{An integer scalar to set the seed. Default to 1.}

\item{verbose}{A logical scalar. Default to TRUE.}
}
\value{
A \code{\link[DNAcopy]{DNAcopy}} object.
}
\description{
Use Conumee's settings of \code{\link[DNAcopy]{segment}}. Notice: Conumee
uses the bins of probes as markers rather than probes themselves.
}
