% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAcopy.R
\name{dnacopy_analysis}
\alias{dnacopy_analysis}
\title{Run DNAcopy analysis.}
\usage{
dnacopy_analysis(x, lrr, seed = 1, verbose = TRUE, ...)
}
\arguments{
\item{x}{A \code{minfi} object such as
\code{\link[minfi]{RGChannelSet-class}} or
\code{\link[minfi]{MethylSet-class}} or
\code{\link[minfi]{GenomicMethylSet-class}} or
\code{\link[minfi]{GenomicRatioSet-class}}.}

\item{lrr}{A matrix of log2 transformed ratio of query and reference copy
number (sum of methylation and unmethylation signals).}

\item{seed}{An integer scalar to set the seed. Default to 1.}

\item{verbose}{A logical scalar. Default to TRUE.}

\item{...}{Arugments passed to \code{\link[DNAcopy]{segment}}.}
}
\value{
A list of two element:
  \itemize{
    \item \code{DNAcopy} a \code{\link[DNAcopy]{DNAcopy}} object.
    \item \code{CNA} a code{\link[DNAcopy]{CNA}} object (not smoothed).
  }
}
\description{
Run DNAcopy analysis.
}
\details{
The function provides a template of run DNAcopy analysis. There are
  many parameters in smoothing and segmentation. The function uses the
  default settings for smoothing.
}
