% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{get_influential_loci}
\alias{get_influential_loci}
\title{Get influential (i.e. outlier) loci}
\usage{
get_influential_loci(x, threshold = 4, verbose = TRUE)
}
\arguments{
\item{x}{An object of \code{\link[minfi]{MethylSet-class}} or
\code{\link[minfi]{GenomicMethylSet-class}}.}

\item{threshold}{A numeric scalar to define outlier. It is the times
that Cook's distance greater than the mean.}

\item{verbose}{A logical scalar.}
}
\value{
A character vector of loci names.
}
\description{
Get influential (i.e. outlier) loci
}
