% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylcnv-pipe.R
\name{methylcnv_pipe}
\alias{methylcnv_pipe}
\title{MethylCNV pipeline.}
\usage{
methylcnv_pipe(ref, qry, report_dir, norm_method = c("methylcnv", "swan",
  "illumina", "raw", "quantile", "noob", "funnorm", "yamat", "dkfz"),
  batch = NULL, batch2 = NULL, pipe_file = "methylCNV.Rda",
  overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{ref}{Reference samples stored in an object of
\code{\link[minfi]{RGChannelSet-class}}.}

\item{qry}{Query samples stored in an object of
\code{\link[minfi]{RGChannelSet-class}}.}

\item{report_dir}{A character scalar of reporting.}

\item{norm_method}{A character scalar of method, including raw, illumina,
swan, quantile, noob, funnorm, yamat, dkfz, quantile. Default to
"methylcnv".}

\item{batch}{factor or vector indicating batches. Default to \code{NULL}, -
do not remove batch effect.}

\item{batch2}{optional factor or vector indicating a second series of
batches. Default to \code{NULL}, - do not remove batch effect.}

\item{pipe_file}{A character scalar of the filename storing the pipeline
result sample by sample. Default to "methylCNV.Rda". Turn off the saving by
set it to NULL.}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}
}
\value{
An object of \code{\link{CMethylCNVPipe}} class.
}
\description{
The method was described in the paper Feng, G. A Statistical Method to
Estimate DNA Copy Number from Illumina High-Density Methylation Arrays.
Systems Biomedicine (2013).
}
\details{
The \code{ref} and \code{qry} should include at least the following
  columns \code{Sample_Name}, \code{Gender}, \code{Batch}, \code{Sample_Prep}.
  Batch effects are removed on methylation and unmethylation signals (log2
  transformed) separately.
}
\note{
The method was named methylCNV according to Feng's paper. It is supposed
  to be implemented in the Lumi package. However, I did not find it by
  searching the name or CNV in its manual.
}
