% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-classes.R
\docType{class}
\name{Pipe-class}
\alias{Pipe-class}
\alias{Pipe}
\title{A S4 class represent CNV pipeline.}
\description{
A S4 class represent CNV pipeline.
}
\section{Slots}{

\describe{
\item{\code{dat}}{An environment which contains data.}

\item{\code{meta}}{A list contain meta information.}
}}

