% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess}
\usage{
preprocess(ref, qry, norm_method = c("swan", "illumina", "raw",
  "quantile", "noob", "funnorm", "yamat", "dkfz"), batch = NULL,
  batch2 = NULL, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{ref}{Reference samples stored in an object of
\code{\link[minfi]{RGChannelSet-class}}.}

\item{qry}{Query samples stored in an object of
\code{\link[minfi]{RGChannelSet-class}}.}

\item{norm_method}{A character scalar of method, including raw, illumina,
swan, quantile, noob, funnorm, yamat, dkfz, quantile. Default to "swan".}

\item{batch}{factor or vector indicating batches. Default to \code{NULL}, -
do not remove batch effect.}

\item{batch2}{optional factor or vector indicating a second series of
batches. Default to \code{NULL}, - do not remove batch effect.}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}
}
\value{
An object of \code{\link[minfi]{GenomicMethylSet-class}} or
  \code{\link[minfi]{MethylSet-class}}. If \code{norm_method} is quantile,
  an object of \code{\link[minfi]{GenomicRatioSet-class}} is returned.
}
\description{
Preprocessing include normalization and batch effect removal.
}
\details{
The \code{ref} and \code{qry} should include at least the following
  columns \code{Sample_Name}, \code{Gender}, \code{Batch}, \code{Sample_Prep}.
  Batch effects are removed on methylation and unmethylation signals (log2
  transformed) separately.
}
