% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\docType{methods}
\name{report_pipe}
\alias{report_pipe}
\alias{report_pipe,CwobPipe-method}
\title{Report an object of pipeline classes.}
\usage{
report_pipe(x, outdir, filename, overwrite = FALSE, verbose = TRUE,
  ...)

\S4method{report_pipe}{CwobPipe}(x, outdir, overwrite = FALSE,
  verbose = TRUE)
}
\arguments{
\item{x}{An object.}

\item{outdir}{A character scalar of output directory.}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}

\item{...}{Any other arguments.}
}
\value{
A character vector of file names.
}
\description{
Report an object of pipeline classes.
}
\section{Methods (by class)}{
\itemize{
\item \code{CwobPipe}: x is \code{CwobPipe}.
}}

