% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\docType{methods}
\name{save_pipe}
\alias{save_pipe}
\alias{save_pipe,ConumeePipe-method}
\title{Save an object of pipeline classes.}
\usage{
save_pipe(x, outdir, filename, overwrite = FALSE, verbose = TRUE, ...)

\S4method{save_pipe}{ConumeePipe}(x, outdir, filename = "conumee.Rda",
  overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{An object.}

\item{outdir}{A character scalar of output directory.}

\item{filename}{A character scalar of output file name.}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}

\item{...}{Any other arguments.}
}
\value{
A character vector of file names.
}
\description{
Save an object of pipeline classes.
}
\section{Methods (by class)}{
\itemize{
\item \code{ConumeePipe}: x is \code{ConumeePipe}.
}}

