% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAcopy.R
\name{summarize_dnacopy_segments}
\alias{summarize_dnacopy_segments}
\title{Summarize \code{DNAcopy} segments.}
\usage{
summarize_dnacopy_segments(x, ...)
}
\arguments{
\item{x}{A \code{\link[DNAcopy]{DNAcopy}} object.}

\item{...}{Arugments passed to \code{\link[DNAcopy]{segment.p}}.}
}
\value{
A \code{data.frame}.
}
\description{
Summarize \code{DNAcopy} segments.
}
