% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conumee-pipe-tune.R
\name{tune_conumee_pipe}
\alias{tune_conumee_pipe}
\title{Run Conumee pipeline with different preprocessing methods.}
\usage{
tune_conumee_pipe(ref, qry, report_dir, norm_methods = c("swan",
  "illumina", "raw", "quantile", "noob", "funnorm", "yamat", "dkfz",
  "methylcnv"), batch = NULL, batch2 = NULL, pipe_file = NULL,
  dnacopy_seed = 1, segment_grid, overwrite = FALSE, verbose = TRUE,
  ...)
}
\arguments{
\item{ref}{Reference samples stored in an object of
\code{\link[minfi]{RGChannelSet-class}}.}

\item{qry}{Query samples stored in an object of
\code{\link[minfi]{RGChannelSet-class}}.}

\item{report_dir}{A character scalar of reporting.}

\item{norm_methods}{A character vector of methods, including raw, illumina,
swan, quantile, noob, funnorm, yamat, dkfz, quantile, methylcnv. Default
to all of them.}

\item{batch}{factor or vector indicating batches. Default to \code{NULL}, -
do not remove batch effect.}

\item{batch2}{optional factor or vector indicating a second series of
batches. Default to \code{NULL}, - do not remove batch effect.}

\item{pipe_file}{A character scalar of the filename storing the pipeline
result sample by sample. Turn off the saving by set it to \code{NULL}.
Default to \code{NULL}.}

\item{dnacopy_seed}{An integer scalar to set the seed. Default to 1.}

\item{segment_grid}{A \code{data.frame} of parameters for \code{\link[conumee]{CNV.segment}}.
If it is not set, generate one use \code{\link{.conumee_segment_grid}}.}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}

\item{...}{Any arguments passed to \code{\link[conumee]{CNV.segment}}.}
}
\value{
An object of \code{\link{ConumeePipe}} class.
}
\description{
Run Conumee pipeline with different preprocessing methods, i.e. the
combination of normalization methods and batch effect correction. The
function serves as a template for writing your own function to run
the pipeline with different parameters, rather than a function of
general purpose.
}
\details{
The \code{ref} and \code{qry} should include at least the following
  columns \code{Sample_Name}, \code{Gender}, \code{Batch}, \code{Sample_Prep}.
  Batch effects are removed on methylation and unmethylation signals (log2
  transformed) separately.
  The function outputs the result in a directory. Each subdirectory is a
  normalzation method. A deeper subdirectory is named in a pattern of
  "\code{batch}-\code{batch2}". Then, it contains the subdirectories of
  sample names, in which there is a segment table named "segments.csv".
}
