% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-segments.R
\docType{methods}
\name{write_segments}
\alias{write_segments}
\alias{write_segments,ConumeePipe-method}
\title{Write segments by sample.}
\usage{
write_segments(x, outdir, filename, overwrite = FALSE, verbose = TRUE,
  ...)

\S4method{write_segments}{ConumeePipe}(x, outdir,
  filename = "conumee-segments.tab", overwrite = FALSE,
  verbose = TRUE)
}
\arguments{
\item{x}{An object.}

\item{outdir}{A character scalar of output directory.}

\item{filename}{A character scalar of output file name.}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}

\item{...}{Any other arguments.}
}
\value{
A character vector of file names.
}
\description{
Write segments by sample.
}
\section{Methods (by class)}{
\itemize{
\item \code{ConumeePipe}: x is \code{ConumeePipe}.
}}

