% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conumee-pipe.R
\name{cn_pipe_conumee}
\alias{cn_pipe_conumee}
\title{CNV pipeline based-upon Conumee package.}
\usage{
cn_pipe_conumee(ref, qry, report_dir, norm_method = c("swan", "illumina",
  "raw", "quantile", "noob", "funnorm", "yamat", "dkfz"), batch = NULL,
  batch2 = NULL, pipe_file = "conumee.Rda", overwrite = FALSE,
  verbose = TRUE)
}
\arguments{
\item{ref}{Reference samples stored in an object of
\code{\link[minfi]{RGChannelSet-class}}.}

\item{qry}{Query samples stored in an object of
\code{\link[minfi]{RGChannelSet-class}}.}

\item{report_dir}{A character scalar of reporting.}

\item{norm_method}{A character scalar of method, including raw, illumina,
swan, quantile, noob, funnorm, yamat, dkfz, quantile. Default to "swan".}

\item{batch}{factor or vector indicating batches. Default to \code{NULL}, -
do not remove batch effect.}

\item{batch2}{optional factor or vector indicating a second series of
batches. Default to \code{NULL}, - do not remove batch effect.}

\item{pipe_file}{A character scalar of the filename storing the pipeline
result sample by sample. Default to "conumee.Rda". Turn off the saving by
set it to NULL.}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}

\item{...}{Any arguments passed to \code{\link[conumee]{CNV.segment}}.}
}
\value{
An object of \code{\link{ConumeePipe}} class.
}
\description{
CNV pipeline based-upon Conumee package.
}
\details{
The \code{ref} and \code{qry} should include at least the following
  columns \code{Sample_Name}, \code{Gender}, \code{Batch}, \code{Sample_Prep}.
  Batch effects are removed on methylation and unmethylation signals (log2
  transformed) separately.
}
