% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-classes.R
\docType{class}
\name{ConumeePipe-class}
\alias{ConumeePipe-class}
\alias{ConumeePipe}
\title{Conumee CNV pipeline}
\description{
Conumee CNV pipeline
}
\section{Slots}{

\describe{
\item{\code{dat}}{An environment which has the following objects:
\itemize{
  \item \code{minfi_obj}: an object of
    \code{\link[minfi]{GenomicRatioSet-class}} or
    \code{\link[minfi]{GenomicMethylSet-class}}.
  \item \code{conumee_results}: a list of \code{\link[conumee]{CNV.analysis-class}}.
}}

\item{\code{meta}}{A list contain meta information.}
}}

