% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\docType{methods}
\name{report_pipe}
\alias{report_pipe}
\alias{report_pipe,CwobPipe-method}
\alias{report_pipe,ConumeePipe-method}
\title{Report an object of pipeline classes.}
\usage{
report_pipe(x, outdir, genome_plot_file = "genome-plot.png",
  genome_plot_width = 9, genome_plot_height = 15, chr_per_row = 4,
  chr_plot_width = 9, chr_plot_height = 6, size = 5e+06,
  cn_boundary = c(1.319508, 2.828427), segment_file = "segments.tab",
  overwrite = FALSE, verbose = TRUE, ...)

\S4method{report_pipe}{CwobPipe}(x, outdir,
  genome_plot_file = "genome-plot-cwob.png", genome_plot_width = 9,
  genome_plot_height = 15, chr_per_row = 4, chr_plot_width = 9,
  chr_plot_height = 6, size = 5e+06, cn_boundary = c(1.319508,
  2.828427), segment_file = "segments-cwob.tab", overwrite = FALSE,
  verbose = TRUE)

\S4method{report_pipe}{ConumeePipe}(x, outdir,
  genome_plot_file = "genome-plot.png", genome_plot_width = 9,
  genome_plot_height = 15, chr_per_row = 4, chr_plot_width = 9,
  chr_plot_height = 6, size = 5e+06, CNscale = "absolute",
  cn_boundary = c(1.319508, 2.828427), segment_file = "segments.tab",
  igv_segment_file = "igv-segments.tab", detail_plot_height = 7,
  detail_plot_width = 5, max_cn = 6, overwrite = FALSE,
  verbose = TRUE)
}
\arguments{
\item{x}{An object.}

\item{outdir}{A character scalar of output directory.}

\item{genome_plot_file}{A character scalar of the genome plot file which
shows all chromosomes.}

\item{genome_plot_width}{An integer scalar. Default to 9.}

\item{genome_plot_height}{An integer scalar. Default to 15.}

\item{chr_per_row}{An integer scalar of the number of chromosomes per row
in genome plot. Default to 4.}

\item{chr_plot_width}{An integer scalar of the width of individual chromosome
plot. Default to 9.}

\item{chr_plot_height}{An integer scalar of the height of individual
chromosome plot. Default to 6.}

\item{size}{An integer scalar of the threshold to define CNVs. Default to
\code{5e6} (5 Mb).}

\item{cn_boundary}{An numeric vector of length two which defines the
boundaries to define CNVs. If a segment has a segment mean lower than
the first element or higher than the second element, it is a CNV. It is
the absolute value, so 2 means no copy number change. Default to
\code{c(1.319508, 2.828427)}.}

\item{segment_file}{A character scalar of segments file which is a table of
segments. Default to "segments-cwob.tab".}

\item{overwrite}{A logical scalar. Default to FALSE.}

\item{verbose}{A logical scalar. Default to TRUE.}

\item{...}{Any other arguments.}

\item{CNscale}{Character string specifying if copy number calls supplied are
relative (i.e.copy neutral == 0) or absolute (i.e. copy neutral ==2). One
of "relative" or "absolute".}

\item{igv_segment_file}{A character scalar of segments file which is a table of
segments. Default to "igv-segments.tab".}

\item{detail_plot_height}{An integer scalar. Default to 7.}

\item{detail_plot_width}{An integer scalar. Default to 5.}

\item{max_cn}{A numeric scalar of the maximum of CN. The CNs are sometime
extremely large. The will remove the data points whose CN are greater than
\code{max_cn}. Default to 6.}
}
\value{
TBD.
}
\description{
Report an object of pipeline classes.
}
\section{Methods (by class)}{
\itemize{
\item \code{CwobPipe}: x is \code{CwobPipe}.

\item \code{ConumeePipe}: x is \code{ConumeePipe}.
}}

